修改农历获取

+++ app/Http/Controllers/Api/CalendarController.php
@@ -15,7 +15,7 @@ class CalendarController extends Controller
     {
         $year = request('year');
         $month = request('month');
-        $day = request('day');
+        $day = (int)request('day');
 
         $lunar = Lunar::fromYmd($year, $month, $day);
         $yi  =  $ji  = $js = $xs = '';
